Remove SONAME for macOS.
This caused an issue I didn't really look much into. However setting
install name causes the compiled object to have a dependency to a shared
lib "." (Not to look in ".", but rather "." *is* a lib).

It makes no sense and is weird, but I can just remove the install name
completely as it doesn't seem to be needed.